+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Added
+- Current group (and ancestors) are now expanded just prior to file save to prevent problem where child is outside of parent group.
+    - https://github.com/StateSmith/StateSmith-drawio-plugin/issues/28
+
+---
+
 ## [0.4.0]
 ### Added
 - UnGroup warning now uses EditorUiHandleError function so that it works with vscode extension as well as browser.